% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_SMC_AuxSP.R
\name{AuxSP.disturb.KM}
\alias{AuxSP.disturb.KM}
\title{Disturbing function used in the bootstrap procedure}
\usage{
AuxSP.disturb.KM(nboot, V.KM, auxinfo)
}
\arguments{
\item{nboot}{specifies the number of bootstrap sampling. The default \code{nboot = 100}.}

\item{V.KM}{the estimated variance-covariance matrix of the auxiliary information based on Kaplan-Meier estimator}

\item{auxinfo}{a matrix that collects the auxiliary information in a pre-specified way.}
}
\description{
Disturb the auxiliary in the bootstrap procedure
}
