% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\docType{package}
\name{CureAuxSP}
\alias{CureAuxSP}
\title{CureAuxSP: Mixture Cure Models with Auxiliary Subgroup Survival Probabilities.}
\description{
This package provides an information synthesis framework that can Estimate mixture cure models with subgroup survival probabilities
as auxiliary information. The underlying methods are based on the paper titled "Efficient auxiliary information synthesis for
cure rate model", which has been published in Jie Ding, Jialiang Li and Xiaoguang Wang (2024) \url{doi:10.1093/jrsssc/qlad106}.
project work
}
