% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{print.SMC.AuxSP}
\alias{print.SMC.AuxSP}
\title{Print SMC.AuxSP object returned by our main function SMC.AuxSP()}
\usage{
\method{print}{SMC.AuxSP}(object)
}
\arguments{
\item{object}{an object of SMC.AuxSP}
}
\description{
Output of \code{SMC.AuxSP} object.
}
