% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Probs.Sub}
\alias{Probs.Sub}
\alias{St.Sub.KM}
\title{Subgroup survival rates}
\usage{
Probs.Sub(tstar, sdata, G)
}
\arguments{
\item{tstar}{time points that the survival rates will be estimated at.}

\item{sdata}{a survival dataset (dataframe) in which to interpret the variables named in the \code{formula} and the \code{cureform}.}

\item{G}{a matrix used to indicate which subgroups he/she belongs to for each of these subjects.}
}
\description{
Calculate Subgroup survival rates basedon the Kaplan-Meier estimation procedure
}
