% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{sdata.SMC}
\alias{sdata.SMC}
\title{Generate simulated dataset from a well-designed PH mixture cure model}
\usage{
sdata.SMC(n)
}
\arguments{
\item{n}{the sample size.}
}
\description{
Generate simulated dataset from a well-designed PH mixture cure model.
}
