% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{SMC.AFT.rank}
\alias{SMC.AFT.rank}
\title{The rank function (with AFT latency)}
\usage{
SMC.AFT.rank(bet, yobs, delta, X, w, intercept = FALSE)
}
\arguments{
\item{bet}{unknown parameters corresponding to latency part, or the fitted values of regression coefficients in the latency part returned by the function \code{\link{SMC.AFT.fit()}}.}

\item{yobs}{time to event of interest.}

\item{delta}{the censoring indicator, normally 1 = event of interest happens, and 0 = censoring.}

\item{X}{a matrix of covariates corresponding to latency part.}

\item{w}{conditional probability of the individual remaining uncured.}

\item{intercept}{a logical value. If \code{TRUE}, the intercept term in the latency part will be included.}
}
\description{
Rank estimating equation used in the M-step of the EM algorithm for the AFT mixture cure model.
}
