% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_Cure.R
\name{SMC.AFT.Sutx}
\alias{SMC.AFT.Sutx}
\title{Survival function for uncured patients (with AFT latency)}
\usage{
SMC.AFT.Sutx(
  yobs,
  delta,
  X,
  bet,
  w,
  cross = FALSE,
  tm = NULL,
  intercept = FALSE
)
}
\arguments{
\item{yobs}{time to event of interest.}

\item{delta}{the censoring indicator, normally 1 = event of interest happens, and 0 = censoring.}

\item{X}{a matrix of covariates corresponding to latency part.}

\item{bet}{unknown parameters corresponding to latency part, or the fitted values of regression coefficients in the latency part returned by the function \code{\link{SMC.AFT.fit()}}.}

\item{w}{conditional probability of the individual remaining uncured.}

\item{cross}{a logical value. The default is \code{FALSE}.}

\item{tm}{the time points that the survival function will be calculated at. The default is \code{NULL}.}

\item{intercept}{a logical value. If \code{TRUE}, the intercept term in the latency part will be included.}

\item{Var}{a logical value. If it is \code{TRUE}, the program returns Standard errors based on the bootstrap method is used. By default, \code{Var = FALSE}.}
}
\description{
Calculate the survival function for uncured patients (with AFT latency) at various time points.
}
