% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_SMC_AuxSP.R
\name{SMC.AuxSP.AIS.fit}
\alias{SMC.AuxSP.AIS.fit}
\title{Semiparametric mixture cure model with auxilary subgrop survival information}
\usage{
SMC.AuxSP.AIS.fit(
  yobs,
  delta,
  X,
  Z,
  aux,
  latency = "ph",
  stdz = FALSE,
  methods.suffix = c("ori", "homo", "hetero_pen", "hetero_ma", "hetero_ma_nestpen",
    "hetero_ma_nestfix"),
  tuning = NULL,
  auxsort = NULL,
  nboot = 100,
  tau.true = NULL
)
}
\arguments{
\item{yobs}{time to event of interest.}

\item{delta}{the censoring indicator, normally 1 = event of interest happens, and 0 = censoring.}

\item{X}{a matrix of covariates corresponding to latency part.}

\item{Z}{a matrix of covariates corresponding to incidence part.}

\item{aux}{indicates the historical summary data. It is a list of lists, and each sub-list represents auxiliary information from a study.
In each study, we have several time points and each time point contains the following four elements
\code{M} the sample size of the external study;
\code{tstar} the time point that the auxiliary information was calculated at;
\code{gfun} a function used to identify the subgroup;
\code{sprob} auxiliary subgroup survival rates for each subgroup at the current time point.}

\item{latency}{specifies the model used in latency part. It can be \code{ph} which represents the proportional hazards model, or \code{aft} which represents the accelerated failure time model.}

\item{stdz}{standardize the covariates. Default is FALSE.}

\item{tuning}{a parameter used in the model averaging auxiliary information systhesis method. Default is log(length(yobs))}

\item{auxsort}{the pre knowledge of the sort of these available auxiliary information. Default is NULL.}

\item{nboot}{specifies the number of bootstrap sampling. The default \code{nboot = 100}.}

\item{method.suffix}{specify the method that will used to accomodate heterogeneous auxiliary information. Note that it is valid only when \code{hetero=TRUE}. The available methods include:
\code{hetero_pen} indicates the penalized auxiliary information synthesis (PAIS) method;
\code{hetero_ma} indicates the model averaging auxiliary information synthesis (MAIS) method (with all possible candidate models);
\code{hetero_ma_nestpen} indicates the model averaging auxiliary information synthesis (MAIS) method (with nested candidate models sorted by PAIS).
\code{hetero_ma_nestfix} indicates the model averaging auxiliary information synthesis (MAIS) method (with nested candidate models defined before).}
}
\description{
Fit the semiparametric mixture cure model with auxilary subgrop survival information.
}
