% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_SMC_AuxSP.R
\name{SMC.AuxSP.AIS.MA.fit}
\alias{SMC.AuxSP.AIS.MA.fit}
\title{Model averaging auxiliary information synthesis method under the mixture cure model}
\usage{
SMC.AuxSP.AIS.MA.fit(indices, bet, gam, dif, pai, W, auxinfo, tuning)
}
\arguments{
\item{indices}{the indices for candidate models (one row indicate one candidate model).}

\item{bet}{unknown parameters corresponding to latency part.}

\item{gam}{unknown parameters corresponding to incidence part.}

\item{dif}{the calculated values of estimating equations related to auxiliary information.}

\item{pai}{the frequencies of each subgroup.}

\item{W}{a matrix}

\item{auxinfo}{a matrix that collects the auxiliary information in a pre-specified way.}

\item{tuning}{a parameter used in the model averaging auxiliary information systhesis method. Default is log(length(yobs))}
}
\description{
Fit the model averaging auxiliary information synthesis method under the mixture cure model
}
