% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_SMC_AuxSP.R
\name{SMC.AuxSP.AIS}
\alias{SMC.AuxSP.AIS}
\title{Semiparametric mixture cure model with auxiliary subgroup survival information using AIS method}
\usage{
SMC.AuxSP.AIS(
  formula,
  cureform,
  data,
  aux,
  latency = c("ph", "aft")[1],
  stdz = FALSE,
  nboot = 100
)
}
\arguments{
\item{formula}{a formula expression, of the form \code{response ~ predictors}.
The \code{response} is a \code{Surv} object with right censoring.
It is used to specify the covariate (risk factor) effects on the failure time of uncured subjects.
See the documentation for \code{survreg} and \code{Surv} in R package \code{survival} for details.
The expression to the right of the "~" specifies the effect of covariates on the failure time of uncured patients.}

\item{cureform}{indicator function a formula expression, of the form \code{cureform ~ predictors}.
It is used to specify the effects of covariates on the cure rate.
A covariate may be used in both \code{formula} and \code{cureform}.}

\item{data}{a data frame in which to interpret the variables named in the \code{formula} and the \code{cureform}.}

\item{aux}{indicates the historical summary data. It is a list of lists, and each sub-list represents auxiliary information from a study.
In each study, we have several time points and each time point contains the following four elements
\code{M} the sample size of the external study;
\code{tstar} the time point that the auxiliary information was calculated at;
\code{gfun} a function used to identify the subgroup;
\code{sprob} auxiliary subgroup survival rates for each subgroup at the current time point.}

\item{latency}{specifies the model used in latency part.
It can be \code{ph} which represents the proportional hazards model, or \code{aft} which represents the accelerated failure time model.}

\item{stdz}{a logical value. If it is \code{TRUE}, all the covariates are standardized. Otherwise, non-transformed covariates are used.}

\item{nboot}{specifies the number of bootstrap sampling. The default \code{nboot = 100}.}
}
\value{
An object of class \code{SMC.AuxSP} is returned. It can be examined by \code{print.SMC.AuxSP()}.
Note that the fitting results of corresponding mixture cure model without taking auxiliary information into consideration will be contained in the returned object.
}
\description{
Fit the semi-parametric mixture cure model with auxiliary subgroup survival probability information using AIS method.
The test statistic for evaluating the homogeneity assumption will also be calculated.
}
\details{
This is a function used to fit the semiparametric mixture cure model with auxilary subgrop survival information.
The method used here is the control variate technique.
The test statistic for evaluating the homogeneity assumption will also be calculated.
}
\examples{
### Example 1. Fit the PHMC model with auxiliary information for the simulated toy data.

set.seed(1)

## split the dataset into two parts: one for internal study and the other for external study
sdata.internal.select <- c(1:5000) \%in\% sample(1:5000,500,replace=FALSE)
sdata.internal <- toydata_smc[sdata.internal.select,]
sdata.external <- toydata_smc[!sdata.internal.select,]

## prepare our auxiliary information: derived from the external study defined above
gfunc = function(X,Z){
  rbind(  ( X[,1] >= -1.0 & X[,1] < -0.5 & X[,2] == 0 ),
          ( X[,1] >= -0.5 & X[,1] <  0.0 & X[,2] == 0 ),
          ( X[,1] >=  0.0 & X[,1] <  0.5 & X[,2] == 0 ),
          ( X[,1] >=  0.5 & X[,1] <  1.0 & X[,2] == 0 ),
          ( X[,1] >= -1.0 & X[,1] < -0.5 & X[,2] == 1 ),
          ( X[,1] >= -0.5 & X[,1] <  0.0 & X[,2] == 1 ),
          ( X[,1] >=  0.0 & X[,1] <  0.5 & X[,2] == 1 ),
          ( X[,1] >=  0.5 & X[,1] <  1.0 & X[,2] == 1 )
  ) * 1
}
sprob <- St.Sub.KM(
  tstar=1,yobs=sdata.external[,1],delta=sdata.external[,2],
  G=gfunc(sdata.external[,-c(1,2)],sdata.external[,-c(1,2)])
)
aux.homo <- list(
  study1 = list(
    time1 = list(
      M = 4500, tstar = 1,
      gfunc = gfunc, sprob = sprob
    )))
aux.hetero <- list(
  study1 = list(
    time1 = list(
      M = 4500, tstar = 1,
      gfunc = gfunc, sprob = sprob + c(0.2,-0.1,0,0,0,0,0,0)
    )))

## fit the model with homogeneous auxiliary information
toydata.SMC.AuxSP.ph <-  SMC.AuxSP.AIS(
  formula = Surv(yobs,delta)~X1+X2, cureform = ~X1, data=sdata.internal,
  aux = aux.homo, latency = "ph", stdz=FALSE, nboot = 100
)
print.SMC.AuxSP(toydata.SMC.AuxSP.ph)

## fit the model with heterogeneous auxiliary information
# without adjusting the heterogeneity (AIS)
toydata.SMC.AuxSP.ph.homo <-  SMC.AuxSP.AIS(
  formula = Surv(yobs,delta)~X1+X2, cureform = ~X1, data=sdata.internal,
  aux = aux.hetero, latency = "ph", stdz=FALSE, nboot = 100
)
print.SMC.AuxSP(toydata.SMC.AuxSP.ph.homo)
# use penalization techniques (PAIS)
toydata.SMC.AuxSP.ph.hetero_pen <-  SMC.AuxSP.PAIS(
  formula = Surv(yobs,delta)~X1+X2, cureform = ~X1, data=sdata.internal,
  aux = aux.hetero, latency = "ph", stdz=FALSE, nboot = 100
)
print.SMC.AuxSP(toydata.SMC.AuxSP.ph.hetero_pen)

}
