% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_SMC_AuxSP.R
\name{SMC.AuxSP.EE.PH}
\alias{SMC.AuxSP.EE.PH}
\title{Estimating equation for the auxiliary subgroup survival rates under the proportional hazards mixture cure model}
\usage{
SMC.AuxSP.EE.PH(bet, gam, yobs, delta, X, Z, w, auxinfo)
}
\arguments{
\item{bet}{unknown parameters corresponding to latency part.}

\item{gam}{unknown parameters corresponding to incidence part.}

\item{yobs}{time to event of interest.}

\item{delta}{the censoring indicator, normally 1 = event of interest happens, and 0 = censoring.}

\item{X}{a matrix of covariates corresponding to latency part.}

\item{Z}{a matrix of covariates corresponding to incidence part.}

\item{w}{conditional probability of the individual remaining uncured.}

\item{auxinfo}{a matrix that collects the auxiliary information in a pre-specified way.}
}
\description{
Calculate the estimating equation for the auxiliary subgroup survival rates under the proportional hazards mixture cure model
}
