% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{SMC.AuxSP.fit}
\alias{SMC.AuxSP.fit}
\title{Semiparametric mixture cure model with auxilary subgrop survival information}
\usage{
SMC.AuxSP.fit(
  yobs,
  delta,
  X,
  Z,
  aux,
  hetero = TRUE,
  N = Inf,
  latency = "PH",
  nboot = 400
)
}
\arguments{
\item{yobs}{time to event of interest.}

\item{delta}{the censoring indicator, normally 1 = event of interest happens, and 0 = censoring.}

\item{X}{a matrix of covariates corresponding to latency part.}

\item{Z}{a matrix of covariates corresponding to incidence part.}

\item{aux}{indicates the historical aggregated statistics. It is a list of lists, and each sub-list represents auxiliary information from the same time point.
We combine multiple time points together and each time point contains the following four elements
\code{tstar} the time point that the auxiliary information was calculated at;
\code{sprob} auxiliary subgroup survival rates for each subgroup at the current time point;
\code{gfunc} a function used to identify the subgroup.}

\item{hetero}{denotes a logical value. If it is \code{TRUE}, the penalization wil be applied to identify the
potential heterogeneous auxiliary subgroup survival rates and make a refinement to the final estimator.}

\item{N}{records the sample size of the external sdata that we extract auxiliary information from.
The default value is \code{N = Inf}, which is the case that the uncertainty is ignored.
If \code{N} is not \code{Inf}, the method that takes the uncertainty into consideration will be adopted automatically.}

\item{latency}{specifies the model used in latency part.
It can be \code{PH} which represents the proportional hazards model, or \code{AFT} which represents the accelerated failure time model.
The default is the PH mixture cure model, that is, \code{latency = "PH"}}

\item{nboot}{specifies the number of bootstrap sampling. The default \code{nboot = 400}.}
}
\description{
Fit the semiparametric mixture cure model with auxiliary subgroup survival information.
}
