% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_SMC_AuxSP.R
\name{SMC.AuxSP.MAIS}
\alias{SMC.AuxSP.MAIS}
\title{Semiparametric mixture cure model with auxiliary subgroup survival information using MAIS method}
\usage{
SMC.AuxSP.MAIS(
  formula,
  cureform,
  data,
  aux = NULL,
  latency = c("ph", "aft")[1],
  tuning = NULL,
  auxsort = NULL,
  stdz = FALSE,
  nboot = 100
)
}
\arguments{
\item{formula}{a formula expression, of the form \code{response ~ predictors}.
The \code{response} is a \code{Surv} object with right censoring.
It is used to specify the covariate (risk factor) effects on the failure time of uncured subjects.
See the documentation for \code{survreg} and \code{Surv} in R package \code{survival} for details.
The expression to the right of the "~" specifies the effect of covariates on the failure time of uncured patients.}

\item{cureform}{indicator function a formula expression, of the form \code{cureform ~ predictors}.
It is used to specify the effects of covariates on the cure rate.
A covariate may be used in both \code{formula} and \code{cureform}.}

\item{data}{a data frame in which to interpret the variables named in the \code{formula} and the \code{cureform}.}

\item{aux}{indicates the historical summary data. It is a list of lists, and each sub-list represents auxiliary information from a study.
In each study, we have several time points and each time point contains the following four elements
\code{M} the sample size of the external study;
\code{tstar} the time point that the auxiliary information was calculated at;
\code{gfun} a function used to identify the subgroup;
\code{sprob} auxiliary subgroup survival rates for each subgroup at the current time point.}

\item{latency}{specifies the model used in latency part.
It can be \code{ph} which represents the proportional hazards model, or \code{aft} which represents the accelerated failure time model.}

\item{stdz}{a logical value. If it is \code{TRUE}, all the covariates are standardized. Otherwise, non-transformed covariates are used.}

\item{nboot}{specifies the number of bootstrap sampling. The default \code{nboot = 100}.}

\item{method}{specify the method that will used to accomodate heterogeneous auxiliary information.
The available methods include:
\code{MAIS} indicates the model averaging auxiliary information synthesis (MAIS) method (with all possible candidate models);
\code{MAIS_Nest} indicates the model averaging auxiliary information synthesis (MAIS) method (with nested candidate models sorted by PAIS).}
}
\value{
An object of class \code{SMC.AuxSP} is returned. It can be examined by \code{print.SMC.AuxSP()}.
Note that the fitting results of corresponding mixture cure model without taking auxiliary information into consideration will be contained in the returned object.
}
\description{
Fit the semi-parametric mixture cure model with heterogeneous auxiliary subgroup survival probability information using PAIS method.
The test statistic for evaluating the homogeneity assumption will also be calculated.
}
\details{
This is a function used to fit the semiparametric mixture cure model with auxilary subgrop survival information.
The method used here is the control variate technique.
The test statistic for evaluating the homogeneity assumption will also be calculated.
}
\examples{
  See examples presented in \code{?SMC.AuxSP.AIS}

}
