% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{SMC.AuxSP.Quad.Pen}
\alias{SMC.AuxSP.Quad.Pen}
\title{Penalized auxiliary information synthesis method under the mixture cure model}
\usage{
SMC.AuxSP.Quad.Pen(bet, gam, dif, pai, V.inv.root, Gam, V.inv, auxinfo)
}
\arguments{
\item{bet}{unknown parameters corresponding to latency part.}

\item{gam}{unknown parameters corresponding to incidence part.}

\item{dif}{the calculated values of estimating equations related to auxiliary information.}

\item{pai}{the frequencies of each subgroup.}

\item{V.inv.root}{the square root of the inverse of the variance-covariance matrix of the estimated coefficients based on the internal data only}

\item{Gam}{correlation between the estimated coefficients based on the internal data only and the estimating equations related to auxiliary information.}

\item{V.inv}{the inverse of the variance-covariance matrix of the estimated coefficients based on the internal data only}

\item{auxinfo}{a matrix that collects the auxiliary information in a pre-specified way.}

\item{indices}{the indices for candidate models (one row indicate one candidate model).}
}
\description{
Fit the penalized auxiliary information synthesis method under the mixture cure model
}
