% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_Cure.R
\name{SMC.PH.fit}
\alias{SMC.PH.fit}
\title{Semiparametric proportional hazards mixture cure model}
\usage{
SMC.PH.fit(
  yobs,
  delta,
  X,
  Z,
  incidence = c("logit"),
  Var = FALSE,
  nboot = 100,
  em.maxit = 100,
  em.tol = 1e-05
)
}
\arguments{
\item{yobs}{time to event of interest.}

\item{delta}{the censoring indicator, normally 1 = event of interest happens, and 0 = censoring.}

\item{X}{a matrix of covariates corresponding to latency part.}

\item{Z}{a matrix of covariates corresponding to incidence part.}

\item{incidence}{specifies the link in incidence part. The "logit", "probit" or complementary loglog ("cloglog") links are available. By default incidence = c("logit").}

\item{Var}{a logical value. If it is \code{TRUE}, the program returns Standard errors based on the bootstrap method is used. By default, \code{Var = FALSE}.}

\item{nboot}{specifies the number of bootstrap sampling. The default \code{nboot = 100}.}

\item{em.maxit}{specifies the maximum iteration number. If the convergence criterion is not met, the EM iteration will be stopped after emmax iterations and the estimates will be based on the last maximum likelihood iteration. The default \code{em.maxit = 100}.}

\item{em.tol}{tolerance for convergence. The default is \code{em.tol = 1e-5}. Iteration stops once the relative change in deviance is less than \code{em.tol}.}
}
\description{
Fit semiparametric proportional hazards (PH) mixture cure model by the EM algorithm.
}
