% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{SMC.PH.Sut}
\alias{SMC.PH.Sut}
\title{Survival function for uncured patients (with PH latency)}
\usage{
SMC.PH.Sut(yobs, delta, X, bet, w, tm = NULL)
}
\arguments{
\item{yobs}{time to event of interest.}

\item{delta}{the censoring indicator, normally 1 = event of interest happens, and 0 = censoring.}

\item{X}{a matrix of covariates corresponding to latency part.}

\item{bet}{unknown parameters corresponding to latency part, or the fitted values of regression coefficients in the latency part returned by the function \code{\link{SMC.PH.fit()}}.}

\item{w}{conditional probability of the individual remaining uncured.}

\item{tm}{the time points that the survival function will be calculated at. The default is \code{NULL}.}

\item{Var}{a logical value. If it is \code{TRUE}, the program returns Standard errors based on the bootstrap method is used. By default, \code{Var = FALSE}.}

\item{cross}{a logical value. The default is \code{FALSE}.}
}
\description{
Calculate the survival function for uncured patients (with PH latency) at various time points.
}
