% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{St.Sub.KM}
\alias{St.Sub.KM}
\title{Subgroup survival rates}
\usage{
St.Sub.KM(tstar, yobs, delta, G)
}
\arguments{
\item{tstar}{time points that the survival rates will be estimated at.}

\item{yobs}{time to event of interest.}

\item{delta}{the censoring indicator, normally 1 = event of interest happens, and 0 = censoring.}

\item{G}{a matrix used to indicate which subgroups he/she belongs to for each of these subjects.}
}
\description{
Calculate Subgroup survival rates basedon the Kaplan-Meier estimation procedure
}
