% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Func_SMC_AuxSP.R
\name{print.SMC.AuxSP}
\alias{print.SMC.AuxSP}
\title{Print SMC.AuxSP object}
\usage{
\method{print}{SMC.AuxSP}(fit)
}
\arguments{
\item{fit}{an object of SMC.AuxSP}
}
\description{
Output of \code{SMC.AuxSP} object.
}
