% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Print.SMC.AuxSP}
\alias{Print.SMC.AuxSP}
\title{Print SMC.AuxSP object returned by our main function SMC.AuxSP()}
\usage{
Print.SMC.AuxSP(object)
}
\arguments{
\item{object}{an object of SMC.AuxSP}
}
\value{
This function has no return value and is used to print the results in a \code{SMC.AuxSP} class with a better presentation.
}
\description{
Output of \code{SMC.AuxSP} object.
}
