% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{Probs.Sub}
\alias{Probs.Sub}
\title{Calculate subgroup survival probabilities}
\usage{
Probs.Sub(tstar, sdata, G)
}
\arguments{
\item{tstar}{time points that the survival probabilities will be estimated at.}

\item{sdata}{a survival dataset (dataframe) in which to interpret the variables named in the \code{formula} and the \code{cureform}.}

\item{G}{a matrix used to indicate which subgroups he/she belongs to for each of these subjects.}
}
\value{
It returns the estimated subgroup survival probabilities for a given survival dataset.
}
\description{
Calculate Subgroup survival probabilities basedon the Kaplan-Meier estimation procedure
}
