% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{sdata.SMC}
\alias{sdata.SMC}
\title{Generate simulated dataset from a well-designed PH mixture cure model}
\usage{
sdata.SMC(n, trace = FALSE)
}
\arguments{
\item{n}{the sample size of the simulated dataset.}

\item{trace}{a logical value that indicates whether the information about cure rate and censoring rate should be printed.}
}
\value{
It returns the simulated dataset from a well-designed PH mixture cure model.
}
\description{
Generate simulated dataset from a well-designed PH mixture cure model.
}
